Move controllers and restControllers to java.marvellator.controller package
