+++ src/main/java/marvellator/controller/GroupController.java
@@ -1,4 +1,4 @@
-package marvellator;
+package marvellator.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
