+++ src/main/java/marvellator/controller/HomeController.java
@@ -1,4 +1,4 @@
-package marvellator;
+package marvellator.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
