+++ src/main/java/marvellator/controller/apiControllers/GroupApiController.java
@@ -1,4 +1,4 @@
-package marvellator.apiControllers;
+package marvellator.controller.apiControllers;
 
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
